# grading-platform-v2/docs/api/openapi-v2.yaml
openapi: 3.1.0
info:
  title: AI Grading Platform V2 API
  version: 0.2.0
  description: >
    V2-only API contract for extension-app and admin-console.
servers:
  - url: http://localhost:3000
    description: local
paths:
...
  /api/v2/licenses/activate:
    post:
      summary: Activate license code
      parameters:
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseActivateRequest"
      responses:
        "200":
          $ref: "#/components/responses/LicenseActivateResponse"
        "201":
          $ref: "#/components/responses/LicenseActivateResponse"
        "409":
          $ref: "#/components/responses/ApiError"
  /api/v2/licenses/status:
    get:
      summary: Query license status
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
      responses:
        "200":
          $ref: "#/components/responses/LicenseStatusResponse"
...
  /api/v2/exams:
    get:
      summary: List exams
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
      responses:
        "200":
          description: OK
    post:
      summary: Create exam
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamCreateRequest"
      responses:
        "201":
          description: Created
        "409":
          $ref: "#/components/responses/ApiError"
...
  /api/v2/gradings/evaluate:
    get:
      summary: Get quota status
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
      responses:
        "200":
          description: OK
    post:
      summary: Evaluate grading with rubric
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradingEvaluateRequest"
      responses:
        "200":
          description: Evaluated
        "409":
          $ref: "#/components/responses/ApiError"
  /api/v2/records:
    get:
      summary: List records
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
      responses:
        "200":
          description: OK
    post:
      summary: Batch create records (compat endpoint)
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordsBatchRequest"
      responses:
        "201":
          description: Created
    delete:
      summary: Delete records by filter
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - in: query
          name: id
          schema:
            type: string
        - in: query
          name: questionNo
          schema:
            type: string
        - in: query
          name: questionKey
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /api/v2/records/batch:
    post:
      summary: Batch create records
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordsBatchRequest"
      responses:
        "201":
          description: Created
  /api/v2/records/{id}:
    delete:
      summary: Delete record by id
      parameters:
        - $ref: "#/components/parameters/XActivationCode"
        - $ref: "#/components/parameters/XDeviceId"
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
components:
  parameters:
    XActivationCode:
      in: header
      name: x-activation-code
      schema:
        type: string
    XDeviceId:
      in: header
      name: x-device-id
      schema:
        type: string
    IdempotencyKey:
      in: header
      name: idempotency-key
      schema:
        type: string
  responses:
    ApiError:
      description: Standard API error payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorEnvelope"
    LicenseActivateResponse:
      description: Activate result
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LicenseActivateResponse"
    LicenseStatusResponse:
      description: Status result
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LicenseStatusResponse"
  schemas:
    LicenseActivateRequest:
      type: object
      required: [activationCode]
      properties:
        activationCode:
          type: string
        deviceId:
          type: string
...
    ExamCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        date:
          type: string
        subject:
          type: string
        grade:
          type: string
        description:
          type: string
...
    GradingEvaluateRequest:
      type: object
      required: [rubric]
      properties:
        imageBase64:
          type: string
        rubric: {}
        studentName:
          type: string
        questionNo:
          type: string
        questionKey:
          type: string
        examNo:
          type: string
    RecordsBatchRequest:
      type: object
      required: [records]
      properties:
        records:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [score, maxScore]
            properties:
              questionNo:
                type: string
              questionKey:
                type: string
              studentName:
                type: string
              name:
                type: string
              examNo:
                type: string
              score:
                type: number
              maxScore:
                type: number
              comment:
                type: string
              breakdown: {}
              deviceId:
                type: string
    ApiErrorEnvelope:
      type: object
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
    LicenseActivateResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
    LicenseStatusResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]